improve german localization: stylistic changes
These changes are not really important and more a thing of opinion, but could nonetheless improve the translation.

The main change is a bit more consistency, e.g. consistent translation of "label" as the same word each time.
Here I chose "Label", it should be common enough in german to be understood well.
Previously it was translated either as "Label", "Etikett", "Beschriftung" or "Bezeichnung".